Fix issues with checking/checking collapsed item
When item was collapsed, checked/unchecked status was not getting applied
to its children due to child controllers being deleted. This change will
prevent a controller being deleted if the item still exists in the parent.

+++ src/tree-internal.component.ts
@@ -113,12 +113,6 @@ export class TreeInternalComponent implements OnInit, OnChanges, OnDestroy, Afte
       (this.tree as any).firstCheckedFired = true;
       this.treeService.fireNodeChecked(this.tree);
     }
-
-    if (this.tree.parent && this.tree.parent.checked) {
-      setTimeout(() => {
-        this.controller.check();
-      });
-    }
   }
 
   public ngOnInit(): void {
@@ -169,7 +163,7 @@ export class TreeInternalComponent implements OnInit, OnChanges, OnDestroy, Afte
   }
 
   public ngOnDestroy(): void {
-    if (get(this.tree, 'node.id', '')) {
+    if (get(this.tree, 'node.id', '') && !(this.tree.parent && this.tree.parent.children.indexOf(this.tree) > -1)) {
       this.treeService.deleteController(this.tree.node.id);
     }
